Basic data protection

See merge request iola-team/oxwall-integration!79

+++ api/Schema/Resolvers/AvatarMutationResolver.php
@@ -2,7 +2,9 @@
 
 namespace Iola\Api\Schema\Resolvers;
 
+use Iola\Api\Auth\Errors\PermissionError;
 use Iola\Api\Contract\Integration\AvatarRepositoryInterface;
+use Iola\Api\Contract\Schema\ContextInterface;
 use Iola\Api\Schema\CompositeResolver;
 
 class AvatarMutationResolver extends CompositeResolver
@@ -10,7 +12,13 @@ class AvatarMutationResolver extends CompositeResolver
     public function __construct(AvatarRepositoryInterface $avatarRepository)
     {
         parent::__construct([
-            "addUserAvatar" => function($root, $args) use ($avatarRepository) {
+            "addUserAvatar" => function($root, $args, ContextInterface $context) use ($avatarRepository) {
+                $userId = $args["userId"]->getId();
+
+                if ($userId !== $context->getViewer()->getUserId()) {
+                    throw new PermissionError();
+                }
+
                 $avatar = $avatarRepository->addAvatar([
                     "userId" => $args["userId"]->getId(),
                     "file" => $args["file"]
@@ -22,9 +30,13 @@ class AvatarMutationResolver extends CompositeResolver
                 ];
             },
 
-            "deleteUserAvatar" => function($root, $args) use ($avatarRepository) {
-                $realId = $args["id"]->getId();
-                $avatar = $avatarRepository->findByIds([$realId])[$realId];
+            "deleteUserAvatar" => function($root, $args, ContextInterface $context) use ($avatarRepository) {
+                $avatarId = $args["id"]->getId();
+                $avatar = $avatarRepository->findByIds([$avatarId])[$avatarId];
+
+                if ($avatar->userId !== $context->getViewer()->getUserId()) {
+                    throw new PermissionError();
+                }
 
                 $avatarRepository->deleteAvatar([
                     "id" => $avatar->id
